// packages/types/src/data.d.ts
export interface ValidationError {
    /**
     * Field name
     */
    field: string;
    /**
     * Error message
     */
    message: string;
    /**
     * Error code
     */
    code?: string;
}